Implement NBD_CMD_DISC.
There may be other commands in flight, and we don't want issuing this
command to disrupt those.

Therefore to properly handle this we issue it as a normal command.
The server may continue to reply to commands in flight after this, but
we modify the PREPARE_FOR_REPLY state so it can handle when the server
closes the connection cleanly (which we should have done before
anyway).