states: Fix parsing of NBD_INFO_EXPORT replies which broke interop with qemu-nbd.
If an NBD_INFO_* reply came back which was larger than our payload
static buffer, we skipped it, but then we tried to parse the skipped
(all zero) data.  Unfortunately this was parsed as a legitimate
NBD_INFO_EXPORT reply (because the NBD_INFO_EXPORT constant happens to
be 0), but because we didn't read the data at all it didn't make any
sense.

Thanks: Eric Blake.