lib: socket: Remove is_eof hack.
The is_eof hack is problematic for TLS because there's no good concept
of testing the socket for EOF (the closest we can get is to read a TLS
CLOSE record from the remote side).  So removing this simplifies the
forthcoming TLS implementation and probably even makes the state
machine easier to follow.