meta-context: Skip error payload if server lacks meta_context
If we tried NBD_OPT_SET_META_CONTEXT, and the NBD server rejected our
attempt with an error message, then we should capture (or skip) that
message off the wire in order to go on without base:allocation, the
same as if the server had not sent a message.  This requires a new
state.