examples, tests: Don't make repeated calls to poll(2).
If the connection isn't ready, we could still set want_to_send.  But
then would fail to send anything after poll(2) returns because at that
point we test the connection and see it isn't ready.  This causes the
code to essentially busy wait in poll(2) until the connection changes
state.