api: Add nbd_can_meta_context
request_meta_context() only works if it is done before connect,
because it determines what to send to the server during handshake;
therefore, it's return value does NOT reflect what the server actually
supports. Telling the user that nbd_block_status() will skip the
callback for contexts not agreed to by the server is rather vague;
let's add an API for grabbing the results after handshake to get a
better feel for what callbacks to expect. It's also easy to enhance an
existing test to add coverage for this.

Note that because a server may pick a different context id per
connection, and when multiple contexts are returned can even reorder
the set per connection, we have to track a per-connection struct
meta_context list (tracking a single list on the overall handle would
break mapping context ids back to the string we pass to the callback),
so the implementation is a bit trickier than other flags where a
common h->eflags should be common among all connections.  Thankfully,
our design intentionally hides the server's context id(s) as an
internal detail not exposed to the client.