handle: Do not allow request_meta_context after connect
With multi-conn, we assume that the server gives us matching
connections (whatever export size and features it advertised on one
connection, all the others should be the same). For features
controlled by the server, we could add code to actually check that
this is true (maybe another day), but in the meantime our assumptions
are not too bad. But for features controlled by the client, we are
setting ourselves up for awkward failures if we allow configuration
changes after the first connection but before the second.  Right now,
request_meta_context is the only client-controlled input that affects
what features we will request from the server, so make that code
explicitly fail once any connection has moved out of the created
state.  The restriction works regardless of whether set_multi_conn is
called before or after request_meta_context.