tests: Fix race in AIO parallel tests.
If the timer expires and there happen to be no commands in flight,
then we can still enter the poll(2) with just events=POLLIN which will
never be satisfied.  The solution is to re-test the loop condition
after setting ‘expired’.

This especially happens during the slower TLS-based test and was
fairly readily reproducible, especially (but not exclusively) when
running under valgrind.