generator: Generate Python documentation.
For Python (or any non-POD non-Perl situation) we need to convert the
POD documentation to plain text.  This requires the pod2text cache
logic we have used elsewhere.

Note this makes the generator slow the first time it is run (where it
is running the slow external command pod2text) but after it has been
run once the result is cached and it will run quickly after that.