api: Add nbd_aio_connect_unix call.
This change adds a new AIO call nbd_aio_connect_unix.  Previously if
you wanted to connect to a Unix domain socket asynchronously you had
to resolve the address into a sockaddr_un yourself and call
nbd_aio_connect.

This is a no-op for existing callers, but simplifies callers that
want to asynchronously connect to a Unix domain socket.