api: Rename nbd_request_meta_context -> nbd_add_meta_context.
Let's be consistent in our naming of APIs.

If an API represents a setting, we have consistently used ‘nbd_set_*’
and ‘nbd_get_*’.

If an API represents a list, we should use ‘nbd_add_*’, ‘nbd_remove_*’
and ‘nbd_clear_*’ (clear the whole list).

By renaming this call while the API is unstable we move towards a more
consistent API in the end.

As well as renaming the API I reworked the documentation to make it
clearer.