api: Add flags parameter to pread, trim, cache, disconnect
This is an API/ABI break, but we have not yet declared stable API.

The NBD spec allows clients to send FLAG_FUA on any command, even
though it recommends that FUA only be sent for write-like commands;
however, without strong reason, I don't see the point in letting
libnbd allow clients to send FLAG_FUA on non-write commands. The NBD
spec also adds the DF flag for reads from a server that supports
structured replies (which an upcoming commit will address), and may
add future flags.

Thinking towards the future, we normally want to refuse a flag that we
know the spec does not permit (for example, we already reject
nbd_pwrite(...,LIBNBD_CMD_FLAG_FUA) if the server did not advertise
FUA support), and the NBD spec tends to not allow new flags without
some form of handshaking to agree that both sides understand the new
flag. But we may also want to add a future API
nbd_set_experimental(nbd, true) as a way to allow a client to pass in
any flags for rapid prototyping against a server, without having to
wait for a new libnbd release that understands the new flag.

As such, all of our transmission-phase APIs should have a flags
parameter, even if callers must always set it to 0 for now.  Note that
in the case of disconnect, nbd_aio_disconnect gets the flag, but
nbd_shutdown remains a general purpose wrapper that works even when
not in transmission phase (that is, there is a reason why we did not
add a mere nbd_disconnect sync wrapper), so it does not need to get a
flag (if you need fine-grained control over shutdown, aio is
sufficient).

We may also want to consider teaching the generator about optional
parameters, and have flags default to 0 if not present in languages
that support that (python being the obvious candidate); sadly, C
varargs are insufficient for our purposes, as we'd have to have some
way to know whether or not to expect to parse a flags argument from a
va_list argument.