generator: Implement unlocked API calls.
There was a "dormant" (ie. never implemented) feature of the generator
for calls which don't need to take the handle lock.  This implements
it.

By making h->state be an atomic field (which should be nearly free on
most normal architectures) we can modify calls which only read
h->state and do nothing else with the handle so they never acquire the
lock.