states: Disable Nagle's algorithm on TCP connections.
We generally want the NBD server to receive our commands as soon as
possible so it can start processing.  Also NBD is likely to be used
mostly on LANs.  So don't delay sending packets.

Note this patch will cause libnbd to send lots of small packets since
we tend to call send(2) often on small buffers.

We can make this setting configurable later if it causes problems, or
explore the TCP_CORK socket option.

This has a dramatic effect on performance.  Using

  $ time nbdkit memory 1M \
         --run 'examples/threaded-reads-and-writes localhost 10809'

with the example adjusted to make NR_CYCLES larger, the time drops
from nearly 4 minutes down to 7 seconds (sic).