api: Generate an additional state check before entering a function.
This change adds a generated check on the current state before we
allow a function to be called.

The check is optional and only handles simple cases such as trying to
connect a handle which is already connected, or issue commands on an
unconnected handle.  (For more complex cases you'd still want to write
the checks by hand).

It also generates documentation automatically, and thus removes
previous hand-written documentation.

It is possible that some of the permissions introduced in this check
are too limiting, but we can make them more permissive in future.