examples: Allow URI in examples taking hostname/port
It is extremely unlikely that someone would want to create a directory
named ./nbd+unix:/, which POSIX rules say would be the resolution of
opening 'nbd+unix:///name?socket=file' as a socket name.  Thus, it is
fairly easy to detect if a lone parameter is intended as a socket name
or as a URI.  In fact, this allows shortening './test localhost 10809'
to the more succinct './test nbd://' thanks to URI defaults.

Not all of our examples were set to support hostname/port, so I only
altered the ones that already did so.