api: nbd_get_version, nbd_supports_uri and nbd_get_package_name.
nbd_get_version returns the library version as a string.

nbd_supports_uri returns whether or not the library was compiled with
NBD URI support (ie. with libxml2).

nbd_get_package_name is fairly useless as it always returns the string
"libnbd", however it replaces a function that was written for the
Python bindings.

These take a handle parameter but don't need to use it.  Changing the
generator to support a whole new class of API calls which don't need a
handle is a massive pain.