lib: Atomically update h->state when leaving the locked region.
Split h->state into:

 - h->public_state = the state on entry to the locked region
   This is also the atomically, publicly visible state.

 - h->state = the real current state of the handle

When we leave the locked region we update h->public_state with
h->state, so that from outside the lock the handle appears to move
atomically from its previous state to the final state without going
through any intermediate states.

Some calls to ‘get_state’ become calls to ‘get_next_state’ if they
need the real state.  Others which need to see the publicly visible
state are changed to ‘get_public_state’.

All calls to ‘set_state’ become ‘set_next_state’ because that is the
real state that gets updated.

The purpose of this patch is to make it easier to reason about the
state in lockless code.