api: Recover from block status callback failure
If the user's extent_fn callback fails, we want to fail the overall
block status command with the same errno, but not give up on the
server.  Do this by skipping any further context chunks from the
server, and relying on the previous patch's ability to preserve the
first error encountered.

Update the interop test with qemu's dirty bitmap context to provoke
callback failure, as well as prove that the connection is still valid.

I intentionally chose to not go as far as documenting EPROTO as our
default errno if the callback returns -1 with errno unchanged,
although it's a one-line modification to dirty-bitmap.c to test that
the default did work.