docs: More nbd_connect_uri cross-references
This augments what is in d20860c6, and may be more useful to a reader
that looks at specific API before the overview.

Also, be clear that we are not yet supporting:

nbds://host/export?tls-type=psk&tls-psk-file=/path/to/file

or similar; until we add that support, for now you'll have to manually
call nbd_set_tls_* before nbd_connect_uri.