uri: Saner parsing of query parameter
Our existing nbd+unix:/// parser used an extreme hack of just doing a
strstr() search for 'socket=' - and as a result, would attempt to
connect to the incorrect socket name if someone used '?mysocket=' or
if there was any subsequent ';' or '&' (query delimitiers) or '%' (a
hex-encoded character) in the rest of the raw query substring of the
overall URI.  Since libxml2 doesn't have a way to parse out individual
queries, copy heavily from libvirt's viruri.c wrapper around libxml2
to pull out the correct name/value pairs from the raw query. This also
opens the door for future patches to honor tls_* queries according to
the (still-being-developed) NBD URI specification.