states: Use MSG_MORE to coalesce messages into single packets.
Since we disabled Nagle's algorithm we may send very small packets
over the wire in some situations where we are calling send(2) from
states that are responsible for small parts of the protocol.  By
setting the MSG_MORE flag we can indicate to the kernel that more data
will follow (usually) immediately and so it can append the data to the
same outgoing packet.

Although there is some variability in the test there is a measurable
benefit.  Using this test:

  $ time nbdkit memory 100M --run 'examples/threaded-reads-and-writes localhost 10809'

before applying this patch:

  real	0m54.151s
  real	0m54.950s
  real	0m55.927s

and after applying this patch:

  real	0m39.154s
  real	0m44.249s
  real	0m44.027s

Thanks: Eric Blake for suggesting this change.