states: Another use for MSG_MORE
Following up to cf1a3045, if we know that we have more requests to
transmit (because the user was queueing up requests while we were busy
elsewhere), and our current request is short (a non-write, or a write
with a small payload), then our current command can be batched with
the next command.

The numbers here were not as dramatic and may be merely in the noise;
over three runs of:

 $ time ~/nbdkit/nbdkit memory 100M --run 'examples/threaded-reads-and-writes localhost 10809'

my machine improved runtime from 12.68s to 12.59s.