examples: Enhance access patterns of threaded-reads-and-writes
A typical client will probably be interleaving small and large
requests, rather than using uniform requests everywhere. Update the
test to randomly simulate this.

For me, this cuts test runtime from 23.1s to 12.6s.  Not quite a 50%
reduction in speed, but close and matches the fact that 50% of the
traffic is now short.