pread: Reject server SR read response with no data chunks
The NBD spec requires that a server doing structured reads must not
succeed unless it covers the entire buffer with reply chunks. In the
general case, this requires a lot of bookkeeping to check whether
chunks were non-overlapping and sufficient, as well as non-zero
lnegth, and we'd rather defer such checking to an optional callback
function. But a callback function will only be reached per chunk,
while we still want to fail the overall read if the callback function
was never called because the server erroneously replied with
NBD_REPLY_TYPE_NONE with no other chunks instead of an expected
NBD_REPLY_TYPE_ERROR*. For this specific error case, the bookkeeping
is much simpler - we merely track if we've seen at least one data
chunk.

Note that the NBD spec states that a client doing a 0-length read
request gets unspecified behavior; but that a server must not send a
0-length OFFSET_DATA or OFFSET_HOLE chunk. qemu-nbd implements a
0-length read by returning just NBD_REPLY_TYPE_NONE, so we special
case things to permit that behavior, as well as permitting 0-length
chunks.  If we ever add a strict mode, it would include preventing a
0-length request and diagnosing a 0-length reply, rather than leaving
the burden on a client callback.