states: Consolidate search for current reply's command
No need to have each state recompute which reply is current. This also
consolidates the logic when a reply has an unexpected handle -
previously, we failed for structured (even though the length makes it
easy to recover) and passed for simple (even though there is nothing
on the wire to state if this response is associated with NBD_CMD_READ
if we did not negotiate structured replies, which would allow our
state machine to start parsing arbitrary data as new responses); we're
better off just always moving to DEAD.

Keeping track of the current command also makes it easier for a future
patch to avoid blocking while in a REPLY cycle - at that point, a
reply_cmd set during the READY state will contain everything needed to
correctly resume parsing when more data arrives from the server.  But
for now, we assert that reply_cmd is NULL during READY.