states: Rewrite NBD_REP_INFO parsing
Ensure that the payload was large enough, reject rather than ignore an
incorrect NBD_INFO_EXPORT size, and emit debug messages for the infos
that we are ignoring.  Using a switch statement will also make it
easier for future patches to parse other NBD_INFO that we will
eventually care about.