set-meta-context: Properly truncate server's context name
The 'len' variable is the overall length of the NBD_REP_META_CONTEXT
payload, which includes the 4-byte context_id field. We were using
strndup() on the undoctored length but on the name field starting at
offset 4, which meant we could end up dup'ing up to 4 bytes of garbage
after the values we read from the server. So far, we appear to have
gotten lucky that those four bytes have tended to start with NUL; I
did not audit if there was a way for a malicious server to
purposefully leave non-zero bytes in the tail, but it is probably
possible by replying with a long context name followed by a short
one. On the other hand, the worst that happens is that the garbage
name doesn't match the user's nbd_add_meta_context() request, so we
behave as if the server does not support the context we wanted after
all.

Fixes: d0ec1ee8