states: Add tests for nbd_pread_structured
With a new enough qemu-nbd, we can test that nbd_pread_structured for
sane operation with interop/structured-read.

It is also possible to test that the callback gets used sanely even
for a connection lacking structured replies, by using nbdkit in
tests/oldstyle.

Not done here, because I'm not comfortable enough with OCaml, and
because we don't have any pre-existing ocaml/examples using pread
for doing actual buffer content validation. However, I did experiment
enough to prove that this change:

| diff --git i/ocaml/examples/extents.ml w/ocaml/examples/extents.ml
| index b49f98e..78e6c4d 100644
| --- i/ocaml/examples/extents.ml
| +++ w/ocaml/examples/extents.ml
| @@ -30,4 +30,12 @@ let () =
|            printf "%d:\t%ld\t%s\n" i entries.(i*2) flags
|          done
|        )
| -  )
| +  );
| +
| +  let buf = Bytes.create 512 in
| +  NBD.pread_structured nbd buf 1024_L () (
| +    fun () sub offs err stat ->
| +      let l = Bytes.length sub in
| +      printf "%d %Ld %d %d\n" l offs err stat
| +  );
| +  printf "done\n"

was sufficient to get './run ocaml/examples/extents.bc' to output
the desired '512 1024 0 1', so it looks like the callback is wired
up correctly.