states: Fix math error
When computing the size of NBD_REPLY_TYPE_ERROR_OFFSET, I went the
wrong direction, which only matters for a non-zero server message
length. (It evaded my notice at the time since I'm not aware of any
NBD server that sends this type of error by default; but I hit it
while intentionally making a temporary hack to qemu-nbd to provoke it
- more on that hack in a later patch adding support for
nbd_pread_callback).

Fixes: e4bd477f6a