states: Fix ERROR_OFFSET sanity check
Due to comparison against the wrong field, the sanity check for the
offset tail field in NBD_REPLY_TYPE_ERROR_OFFSET could never fire,
unless a server sent a non-standard NBD_E* code that coincides with
the reply type.  While at it, sink the setting of cmd->error after the
sanity checking. (The bug evaded my notice at the time since I'm not
aware of any NBD server that sends this type of error by default; but
I hit it while intentionally making a temporary hack to qemu-nbd to
provoke it - more on that hack in a later patch adding support for
nbd_pread_callback).

Fixes: 6f238add6