states: Avoid wasted send() when REPLY interrupts request
When we are blocked waiting for POLLOUT during a request, and happen
to receive notice of POLLIN instead, we know that the work done in
response to POLLIN will be non-blocking (thanks to the previous patch,
it returns to %.READY as soon as it would block, which in turn jumps
right back into ISSUE_COMMAND because we have a pending request not
fully sent yet).  Since the jaunt through REPLY was non-blocking, it
is unlikely that the POLLOUT situation has changed in the meantime, so
if we use SET_NEXT_STATE() to step back into SEND_REQUEST with
execution of the corresponding C code, our send() call will likely
fail with EAGAIN, once again blocking us until our next POLLOUT.
Although the wasted syscall is not on the hot-path when we have
already saturated the server with as many requests as possible (after
all, we can't progress until the server is ready), it's slightly
cleaner if we instead declare that we are already blocked.  There may
be other corner cases where performance improvements are more
noticeable.

I tested with:
$ nbdkit -U - null 16M --run 'examples/threaded-reads-and-writes $unixsocket'

There was no real noticeable difference in timing, but I did observe
that pre-patch, the run encountered 168825 pre-emptions and 136976
send() EAGAIN failures (81%), while post-patch the run encountered
166066 pre-emptions and 552 EAGAIN failures (0.3%).