docs: Improve nbd_aio_get_direction documentation
Mention things to remember for avoiding the deadlock of polling for a
POLLIN from a server that has no replies to send.  Also mention that
get_direction CAN return 0.

Perhaps we could split the READY state into two - one for when there
are no commands in flight (and get_direction returns 0 to state that
polling is pointless, although a multi-threaded reader can still poll
for POLLIN), and the other when there ARE commands in flight.  Such a
separation would also make it easier to know when nbd_aio_disconnect
is not going to strand an in-flight command's response. But I'm not
sure it is worth the effort.