aio: Don't duplicate strerror
Calling set_error(value,...) already appends strerror(value) to the
resulting string; no need to repeat it ourself.

Ultimately, we should fix our code to not use strerror(), as it is not
thread-safe. This patch does not do that (because using strerror_r()
is a portability nightmare and strerror_l() is not yet widespread);
but the fewer places where we use strerror(), the more confined that
future patch will be.

Fixes: 1d73554a